chatbot/requirements.txt
streamlit>=1.24.0
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-community>=0.0.8
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pandas>=2.0.0
pyjwt>=2.6.0
requests>=2.31.0
openai>=1.3.0